refactor: formatCommands のテスト重複を解消する (#358)
mc-bridge-minecraft.test.ts の formatCommands テストは仕様テスト相当の
内容であり mc-bridge-format-commands.spec.ts と重複していた。
.test.ts にのみ存在したテストケース（JST変換エッジケース、添付ファイル
詳細、system/bot非エスケープ確認、エラーイベント混在等）を .spec.ts に
移植し、.test.ts を削除した。
